+++ git_acp/modules/git_acp.py
@@ -66,6 +66,20 @@ options:
             - Git repo URL.
         required: True
         type: str
+    remote:
+        description:
+            - Local system alias for git remote PUSH and PULL repository operations.
+        type: str
+        default: origin
+    user_name:
+        decription:
+            - Explicit git local user name. Nice to have for remote operations.
+        type: str
+    user_email:
+        decription:
+            - Explicit git local email address. Nice to have for remote operations.
+        type: str
+
 requirements:
     - git>=2.10.0 (the command line tool)
 '''
@@ -78,6 +92,7 @@ EXAMPLES = '''
     path: /Users/git/git_acp
     branch: master
     comment: Add file1.
+    remote: origin
     add: [ "." ]
     mode: https
     url: "https://gitlab.com/networkAutomation/git_test_module.git"
@@ -88,8 +103,11 @@ EXAMPLES = '''
     branch: master
     comment: Add file1.
     add: [ file1  ]
+    remote: dev_test
     mode: ssh
     url: "git@gitlab.com:networkAutomation/git_test_module.git"
+    user_name: lvrfrc87
+    user_email: lvrfrc87@gmail.com
 
 - name: LOCAL | push on local repo.
   git_acp:
@@ -114,73 +132,156 @@ output:
 from ansible.module_utils.basic import AnsibleModule
 
 
-def git_add(module):
+def user_conifg(module):
 
+    result = dict()
+    user_name = module.params.get('user_name') 
+    user_email = module.params.get('user_email')
+    path = module.params.get('path')
+
+    get_name_cmd = [
+        'git',
+        'config', 
+        '--local', 
+        'user.name',
+    ]
+
+    _rc, output, _error = module.run_command(get_name_cmd, cwd=path)
+    if output != user_name:
+
+        name_cmd = [
+            'git',
+            'config', 
+            '--local', 
+            'user.name',
+            user_name
+        ]
+
+        conf_name = module.run_command(name_cmd, cwd=path)
+        result.update(
+            local_user_name=conf_name,
+            changed=True
+        )
+
+    get_email_cmd = [
+        'git',
+        'config', 
+        '--local', 
+        'user.email',
+    ]
+
+    _rc, output, _error = module.run_command(get_email_cmd, cwd=path)
+    if output != user_email:
+
+        email_cmd = [
+            'git',
+            'config', 
+            '--local', 
+            'user.email',
+            user_email
+        ]
+
+        conf_email = module.run_command(email_cmd, cwd=path)
+        result.update(
+            local_user_email=conf_email,
+            changed=True
+        )
+
+    return result
+
+
+def git_add(module):
+    """
+    Run git add and stage changed files.
+
+    args:
+        * module:
+            type: dict()
+            descrition: Ansible basic module utilities and module arguments.
+    return: null
+    """
     add = module.params.get('add')
     path = module.params.get('path')
 
-    add_cmds = [
+    add_cmd = [
         'git',
         'add',
         '--',
     ]
-    add_cmds.extend(add)
+    add_cmd.extend(add)
 
-    rc, _output, error = module.run_command(add_cmds, cwd=path)
+    rc, _output, error = module.run_command(add_cmd, cwd=path)
 
     if rc != 0:
-        module.fail_json(msg=error)
+        module.fail_json(rc=rc, msg=error, command=' '.join(add_cmd))
     if rc == 0:
         return
 
 
 def git_status(module):
     """
-    Run git status and check if repo has changes
-    :param module:
-    :return: list of files that changes in repo
-    :rtype: set
+    Run git status and check if repo has changes.
+
+    args:
+        * module:
+            type: dict()
+            descrition: Ansible basic module utilities and module arguments.
+    return:
+        * data: 
+            type: set()
+            description: list of files changed in repo.
     """
     data = set()
     path = module.params.get('path')
-    cmd = [
+    status_cmd = [
         'git',
         'status',
         '--porcelain',
     ]
 
-    rc, output, error = module.run_command(cmd, cwd=path)
+    rc, output, error = module.run_command(status_cmd, cwd=path)
 
     if rc != 0:
-        module.fail_json(msg=error)
+        module.fail_json(rc=rc, msg=error, command=' '.join(status_cmd))
     if rc == 0:
-        for i in output.split('\n'):
-            file_name = i.split(' ')[-1].strip()
+        for line in output.split('\n'):
+            file_name = line.split(' ')[-1].strip()
             if file_name:
                 data.add(file_name)
     return data
 
 
 def git_commit(module):
+    """
+    Run git commit and commit files in repo.
 
-    result = dict(changed=False)
+    args:
+        * module:
+            type: dict()
+            descrition: Ansible basic module utilities and module arguments.
+    return:
+        * result: 
+            type: dict()
+            desription: returned output from git commit command and changed status
+    """
+    result = dict()
     comment = module.params.get('comment')
     path = module.params.get('path')
 
     if comment:
         git_add(module)
 
-        commit_cmds = [
+        commit_cmd = [
             'git',
             'commit',
             '-m',
             comment,
         ]
 
-    rc, output, _error = module.run_command(commit_cmds, cwd=path)
+    rc, output, error = module.run_command(commit_cmd, cwd=path)
 
     if rc != 0:
-        module.fail_json(msg=output)
+        module.fail_json(rc=rc, msg=error, command=' '.join(commit_cmd))
     if rc == 0:
         if output:
             result.update(
@@ -191,47 +292,99 @@ def git_commit(module):
 
 
 def git_push(module):
+    """
+    Set URL and remote if required. Push changes to remote repo.
 
+    args:
+        * module:
+            type: dict()
+            descrition: Ansible basic module utilities and module arguments.
+    return:
+        * result: 
+            type: dict()
+            desription: returned output from git push command and updated changed status.
+    """
     def git_set_url(module):
+        """
+        Set URL and remote if required.
 
+        args:
+            * module:
+                type: dict()
+                descrition: Ansible basic module utilities and module arguments.
+        return: null
+        """
         url = module.params.get('url')
         mode = module.params.get('mode')
         user = module.params.get('user')
         token = module.params.get('token')
+        origin = module.params.get('remote')
 
-        if mode == 'https':
-            if url.startswith('https://'):
-                remote_add = [
-                    'git',
-                    'remote',
-                    'set-url',
-                    'origin',
-                    'https://{user}:{token}@{url}'.format(
-                        url=url[8:],
-                        user=user,
-                        token=token,
-                    ),
-                ]
-            else:
-                module.fail_json(msg='HTTPS mode selected but not HTTPS URL provided')
-        else:
-            remote_add = [
-                'git',
-                'remote',
-                'set-url',
-                'origin',
-                url,
-            ]
+        get_url_cmd = [
+            'git',
+            'remote',
+            'get-url',
+            '--all',
+            origin,
+        ]
+        
+        path = module.params.get('path')
 
-        rc, _output, error = module.run_command(remote_add, cwd=path)
+        rc, _output, _error = module.run_command(get_url_cmd, cwd=path)
 
-        if rc != 0:
-            module.fail_json(msg=error)
         if rc == 0:
             return
+        
+        if rc == 128:
+            if mode == 'https':
+                if url.startswith('https://'):
+                    remote_add_cmd = [
+                        'git',
+                        'remote',
+                        'add',
+                        origin,
+                        'https://{user}:{token}@{url}'.format(
+                            url=url[8:],
+                            user=user,
+                            token=token,
+                        ),
+                    ]
+                else:
+                    module.fail_json(msg='HTTPS mode selected but not HTTPS URL provided')
+            else:
+                remote_add_cmd = [
+                    'git',
+                    'remote',
+                    'add',
+                    origin,
+                    url,
+                ]
+
+            rc, _output, error = module.run_command(remote_add_cmd, cwd=path)
+
+            if rc != 0:
+                module.fail_json(msg=error, command=' '.join(remote_add_cmd), rc=rc)
+            if rc == 0:
+                return
+
 
     def git_push_cmd(path, cmd_push):
-        result = dict(changed=False)
+        """
+        Set URL and remote if required. Push changes to remote repo.
+
+        args:
+            * path:
+                type: path
+                descrition: git repo local path.
+            * cmd_push:
+                type: list()
+                descrition: list of commands to perform git push operation.
+        return:
+            * result: 
+                type: dict()
+                desription: returned output from git push command and updated changed status.
+        """
+        result = dict()
 
         rc, output, error = module.run_command(cmd_push, cwd=path)
 
@@ -247,22 +400,23 @@ def git_push(module):
     branch = module.params.get('branch')
     push_option = module.params.get('push_option')
     path = module.params.get('path')
+    origin = module.params.get('remote')
 
-    cmd_push = [
+    push_cmd = [
         'git',
         'push',
-        'origin',
+        origin,
         branch,
     ]
 
     if not push_option:
         git_set_url(module)
-        return git_push_cmd(path, cmd_push)
+        return git_push_cmd(path, push_cmd)
 
     if push_option:
-        cmd_push.insert(3, '--push-option={0} '.format(push_option))
+        push_cmd.insert(3, '--push-option={0} '.format(push_option))
         git_set_url(module)
-        return git_push_cmd(path, cmd_push)
+        return git_push_cmd(path, push_cmd)
 
 
 def main():
@@ -277,20 +431,30 @@ def main():
         push_option=dict(),
         mode=dict(choices=["ssh", "https", "local"], default='ssh'),
         url=dict(required=True),
+        remote=dict(default="origin"),
+        user_name=dict(),
+        user_email=dict(),
     )
 
     required_if = [
         ("mode", "https", ["user", "token"]),
     ]
 
+    required_together = [
+        ["user_name", "user_email"],
+    ]
+
     module = AnsibleModule(
         argument_spec=argument_spec,
         required_if=required_if,
+        required_together=required_together,
     )
 
     url = module.params.get('url')
     push_option = module.params.get('push_option')
     mode = module.params.get('mode')
+    user_name = module.params.get('user_name')
+    user_email = module.params.get('user_email')
 
     if mode == 'local':
         if url.startswith(('https://', 'git@', 'ssh://git@')):
@@ -310,8 +474,13 @@ def main():
         if url.startswith('ssh://git@github.com'):
             module.fail_json(msg='GitHub does not support "ssh://" URL. Please remove it from url: '+url+'')
 
-    result = {'changed': False}
+    result = dict(changed=False)
+
+    if user_name and user_email:
+        result.update(user_conifg(module))
+
     changed_files = git_status(module)
+
     if changed_files:
         result.update(git_commit(module))
         result.update(git_push(module))